requirements.txt
absl-py==0.1.13
asn1crypto==0.24.0
astor==0.6.2
astroid==1.6.3
autopep8==1.3.4
bleach==1.5.0
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
conda==4.4.10
cryptography==2.1.4
cycler==0.10.0
gast==0.2.0
grpcio==1.10.0
html5lib==0.9999999
idna==2.6
isort==4.3.4
Keras==2.1.5
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==2.6.11
matplotlib==2.2.2
mccabe==0.6.1
numpy==1.14.2
pandas==0.22.0
Pillow==5.1.0
protobuf==3.5.2.post1
pycodestyle==2.3.1
pycosat==0.6.3
pycparser==2.18
pylint==1.8.3
pyOpenSSL==17.5.0
pyparsing==2.2.0
PySocks==1.6.7
python-dateutil==2.7.2
pytz==2018.3
PyYAML==3.12
requests==2.18.4
ruamel-yaml==0.15.35
scipy==1.0.1
six==1.11.0
tensorboard==1.7.0
tensorflow==1.7.0
termcolor==1.1.0
urllib3==1.22
Werkzeug==0.14.1
wrapt==1.10.11
